requirements.txt
aiofiles==0.5.0
aniso8601==8.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coverage==5.1
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.54.1
graphene==2.1.8
graphql-core==3.1.0
graphql-relay==3.0.0
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
pydantic==1.5
PyNaCl==1.3.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.23.0
Rx==3.1.0
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.3
ujson==2.0.3
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
